Fix logger crashing after some days/weeks
Only a cold-reset made the sensor available again.

Originally there where two scripts running, both accessing the DHT sensor.
One was reading the sensor and showing its value on an OLED display, the
other was reading the sensor and pushing its value into a round-robin
database (RRDtool).

Now both scripts are combined and the sensor is only accessed from one point.

It seems that the Adafruit driver and/or the Pi does not like it when
two scripts try to access the sensor.